feat(vision): 添加本地离线人脸检测及配套适配
集成乐鑫esp-dl和human_face_detect组件，替换原有esp-sr人脸检测方案
适配16MB Flash分区表，配置2×5MB OTA分区以容纳人脸检测模型
优化ESP32S3摄像头参数：使用CIF分辨率、双缓冲帧、抓取最新完整帧减少画面撕裂
将预览图片显示时长从5秒延长至30秒
添加编译选项关闭Windows工具链的格式不匹配警告，修复多处格式化打印警告
新增本地视觉检测的MCP工具，明确区分本地/远程检测并提示本地检测无数据上传
更新MCP工具文档，提示连续本地检测时无法使用云端拍照工具
修复线程栈过小导致的崩溃问题，适配atk-dnesp32s3开发板的背光控制
移除emote_display编译依赖，添加头文件条件编译逻辑避免缺失头文件报错
更新项目依赖组件版本，修正远程检测器的队列发送错误和日志格式问题
修复显示屏动态转换失败的日志提示问题

+++ main/vision/vision_display.cc
@@ -220,7 +220,7 @@ std::unique_ptr<LvglImage> VisionDisplay::ComposePreview(const uint8_t* frame_rg
     size_t total = dst_stride * (size_t)height;
     uint8_t* dst = (uint8_t*)heap_caps_malloc(total, MALLOC_CAP_SPIRAM);
     if (dst == nullptr) {
-        ESP_LOGE(TAG, "Failed to allocate %zu bytes for preview", total);
+        ESP_LOGE(TAG, "Failed to allocate %u bytes for preview", (unsigned)total);
         return nullptr;
     }
